# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ramekin-server",
    "description": "",
    "contact": {
      "name": "Gregor Stocks"
    },
    "license": {
      "name": "LicenseRef-Proprietary",
      "identifier": "LicenseRef-Proprietary"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/enrich": {
      "post": {
        "tags": [
          "enrich"
        ],
        "summary": "Enrich a recipe using AI",
        "description": "This is a stateless endpoint that takes a recipe object and returns an enriched version.\nIt does NOT modify any database records. The client can apply the enriched data\nvia a normal PUT /api/recipes/{id} call.\n\nCurrently enriches tags by suggesting from the user's existing tag library.",
        "operationId": "enrich_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeContent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Enriched recipe object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeContent"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "AI service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/recipes": {
      "get": {
        "tags": [
          "recipes"
        ],
        "operationId": "list_recipes",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items to return (default: 20, max: 1000)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of items to skip (default: 0)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Search query with optional filters. Supports:\n- Plain text: searches title and description\n- tag:value: filter by tag (can use multiple)\n- source:value: filter by source name\n- has:photos / no:photos: filter by photo presence\n- created:>2024-01-01: created after date\n- created:<2024-12-31: created before date\n- created:2024-01-01..2024-12-31: created in date range\n\nExample: \"chicken tag:dinner tag:quick has:photos\"",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "description": "Sort field (default: updated_at)",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortBy"
            }
          },
          {
            "name": "sort_dir",
            "in": "query",
            "description": "Sort direction (default: desc). Ignored when sort_by=random.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Direction"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of user's recipes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListRecipesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "recipes"
        ],
        "operationId": "create_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Recipe created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRecipeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRecipeRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RecipeContent"
          },
          {
            "type": "object",
            "properties": {
              "photo_ids": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        ]
      },
      "CreateRecipeResponse": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "Direction": {
        "type": "string",
        "description": "Sort direction",
        "enum": [
          "desc",
          "asc"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Shared error response used by all endpoints",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "Ingredient": {
        "type": "object",
        "required": [
          "item"
        ],
        "properties": {
          "amount": {
            "type": [
              "string",
              "null"
            ]
          },
          "item": {
            "type": "string"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "unit": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ListRecipesResponse": {
        "type": "object",
        "required": [
          "recipes",
          "pagination"
        ],
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetadata"
          },
          "recipes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeSummary"
            }
          }
        }
      },
...
      "PaginationMetadata": {
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "limit": {
            "type": "integer",
            "format": "int64",
            "description": "Number of items requested (limit)"
          },
          "offset": {
            "type": "integer",
            "format": "int64",
            "description": "Number of items skipped (offset)"
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of items available"
          }
        }
      },
...
      "RecipeContent": {
        "type": "object",
        "description": "Core recipe content - all fields that can be enriched by AI.\nUsed for: Enrich API request/response, base for CreateRecipeRequest.",
        "required": [
          "title",
          "ingredients",
          "instructions"
        ],
        "properties": {
          "cook_time": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "difficulty": {
            "type": [
              "string",
              "null"
            ]
          },
          "ingredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ingredient"
            }
          },
          "instructions": {
            "type": "string"
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "nutritional_info": {
            "type": [
              "string",
              "null"
            ]
          },
          "prep_time": {
            "type": [
              "string",
              "null"
            ]
          },
          "rating": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "servings": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string"
          },
          "total_time": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "RecipeSummary": {
        "type": "object",
        "required": [
          "id",
          "title",
          "tags",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "thumbnail_photo_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Photo ID of the first photo (thumbnail), if any"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SortBy": {
        "type": "string",
        "description": "Sort field for recipe list",
        "enum": [
          "updated_at",
          "random"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}